packages/templates/src/research-agent/templates/raw-python/requirements.txt
# Core dependencies for Raw Python Research Agent
requests>=2.31.0
beautifulsoup4>=4.12.0
openai>=1.0.0
anthropic>=0.7.0
python-dotenv>=1.0.0

# Optional: Enhanced HTML parsing
lxml>=4.9.0

# Optional: For async support (uncomment if needed)
# aiohttp>=3.8.0